Simplify grade-school\src\lib.rs

+++ grade-school/src/lib.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 #[allow(clippy::new_without_default)]
 pub struct School {
-    roster: HashMap<u32, Vec<String>>,
+    roster: HashMap<u8, Vec<String>>,
 }
 
 impl School {
@@ -10,21 +10,21 @@ impl School {
         Self { roster: HashMap::new() }
     }
 
-    pub fn add(&mut self, grade: u32, student: &str) {
+    pub fn add(&mut self, grade: u8, student: &str) {
         self.roster.entry(grade)
             .or_insert(Vec::new())
             .push(student.to_string());
     }
 
-    pub fn grades(&self) -> Vec<u32> {
-        let mut grades: Vec<u32> = 
+    pub fn grades(&self) -> Vec<u8> {
+        let mut grades: Vec<u8> = 
             self.roster.keys().copied()
-                .collect::<Vec<u32>>();
+                .collect::<Vec<u8>>();
         grades.sort_unstable();
         grades
     }
 
-    pub fn grade(&self, grade: u32) -> Vec<String> {
+    pub fn grade(&self, grade: u8) -> Vec<String> {
         let mut students = 
             self.roster.get(&grade).cloned()
                 .unwrap_or(Vec::new());
